+++ app/src/main/java/com/kopim/productlist/data/model/network/apimodels/addtocart/AddToCartRequestData.kt
@@ -1,6 +1,6 @@
 package com.kopim.productlist.data.model.network.apimodels.addtocart
 
 data class AddToCartRequestData(
-    val list: Int,
+    val list: Long,
     val product: String
 )
